First mattermost notification implementation
This change add ability to send build status notification on mattermost.
It's not perfect, and requires some specs!

To have notification works, you need to configure mattermost In your
~/.bamboo/worker.yml:
```
mattermost:
  url: [URL where bamboo-worker will POST notification]
  username: [Username to display on mattermost]
  icon_url: [The icon url for notification message on mattermost]
  icon_failure: [An image url for failures]
  icon_success: [An image url for success]
```

Only the [mattermost][url] is required, all others are optional but the
mattermost message will be better with these options defined :)